requirements.txt
fastapi
uvicorn
gunicorn
python-dotenv
httpx
google-generativeai
SQLAlchemy
mysql-connector-python
sqlalchemy-utils
python-multipart
moviepy